% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confIntDemo.R
\name{confIntDemo}
\alias{confIntDemo}
\title{Confidence Interval Demonstration}
\usage{
confIntDemo(distr = "normal", size = 20, conf.level = 0.95)
}
\arguments{
\item{distr}{distribution of the population to be sampled. Options include
\code{"normal"}, \code{"exponential"}, \code{"uniform"} and \code{"binary"}
(partial match allowed).}

\item{size}{sample size}

\item{conf.level}{confidence level.}
}
\value{
The command invisibly returns the fraction of intervals that capture
the true mean.
}
\description{
Draw many random samples and compute confidence interval.  How many
intervals capture the true mean?
}
\details{
This simulation will draw 100 random samples from a given population
distribution and compute the correpsonding confidence intervals. The 100
intervals will be drawn with an indication of the ones that missed the true
mean. A histogram of the population will also be created.
}
\examples{

confIntDemo()

confIntDemo(distr = "exponential", size = 40)


}
\author{
Laura Chihara
}
\keyword{confidence}
\keyword{interval}
