% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corDemo.R
\name{corDemo}
\alias{corDemo}
\title{Correlation demonstration}
\usage{
corDemo(r = 0)
}
\arguments{
\item{r}{a number between -1 and 1. Enter any number r, \eqn{latex}{abs(r)>
1 }, to exit the interactive session[}
}
\description{
For a given \code{r}, create a scatterplot of two variables with that
correlation.
}
\details{
Demonstrate the concept of correlation by inputting a number between -1 and
1 and seeing a scatter plot of two variables with that correlation. Once you
invoke this command, you can continue to enter values for r. Type any number
\eqn{latex}{abs(r)> 1 }) to exit.
}
\examples{

\dontrun{
corDemo()
}

}
\author{
Laura Chihara
}
\keyword{Correlation}
