% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Milkshakes}
\alias{Milkshakes}
\title{Milkshakes (chocolate)
Nutrional information on chocolate milkshakes from a sample of restaurants.}
\format{
A data frame with 29 observations on the following 11 variables.
\describe{
\item{Restaurant}{Names of restaurants}
\item{Type}{Type of restaurant, \code{Dine In} \code{Fast Food}}
\item{Calories}{Calories per serving}
\item{Fat}{Fat per
serving (g)}
\item{Sodium}{Sodium per serving (mg)}
\item{Carbs}{Carbohydrates per serving (g)}
\item{SizeOunces}{Size of milkshake (ounces)}
\item{CalPerOunce}{Calories per ounce}
\item{FatPerOunce}{Fat per ounce}
\item{CarbsPerOunce}{Carbohydrates per ounce}
}
}
\source{
Data collected by Carleton students Yoni Blumberg (2013) and Lindsay
Guthrie (2013).
}
\description{
Milkshakes (chocolate)
Nutrional information on chocolate milkshakes from a sample of restaurants.
}
\keyword{datasets}
