% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{penguin_survival}
\alias{penguin_survival}
\title{Penguin survival by tag type}
\format{
A data frame with 2 observations on the following 3 variables.
\describe{
  \item{TagType}{Tag type: \code{Metal} or \code{Electronic}}
  \item{SurStatus}{Survival status: \code{Survived} or \code{Electronic}}
  \item{Died}{Number of penguins that died}
}
}
\source{
Saraux, C., et al., “Reliability of flipper-banded penguins as
indicators of climate change,” Nature, January 2011; 469: 203–206.

Lock, Robin H., et al. Statistics: Unlocking the power of data.
John Wiley & Sons, 2025.
}
\description{
Survival outcomes for penguins fitted with metal or electronic tags.
}
\keyword{datasets}
