% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.carlboot}
\alias{print.carlboot}
\title{Print a summary of an \code{carlboot} object}
\usage{
\method{print}{carlboot}(x, ...)
}
\arguments{
\item{x}{The carlboot object to print.}

\item{...}{not used}
}
\description{
Print summary statistics and confidence intervals, if desired, for an \code{lmeresamp} object.
}
