% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{states03}
\alias{states03}
\title{US government data, 2003}
\format{
A data frame with 50 observations on the following 24 variables.
\describe{
\item{State}{the 50 states}
\item{Region}{a
factor with levels \code{Midwest}, \code{Northeast}, \code{South},
\code{West}}
\item{Pop}{Population in 1000}
\item{Births}{Number of births}
\item{Deaths}{Number of
deaths}
\item{Pop18}{Percent of population 18 years of age or
younger}
\item{Pop65}{Percent of population 65 years of age or
older}
\item{HSGrad}{Percent of population 25 years of age or older
with a high school degree}
\item{ColGrad}{Percent of population 25
years of age or older with a college degree}
\item{TeacherPay}{Average teachers salary in dollars}
\item{InfMortality}{Infant mortality per 1000 live births}
\item{TeenBirths}{Live births per 1000 15-19 year old females}
\item{ViolentCrime}{Violent crime per 100000 population}
\item{PropertyCrime}{Property crime per 100000 population}
\item{DeathPenalty}{State has death penalty?}
\item{Executions}{Number of executions 1977-2003}
\item{Poverty}{Percent of populaton below the poverty level}
\item{Unemp}{Percent unemployed (of population 16 years or older)}
\item{Uninsured}{Percent uninsured (3 year aveage)}
\item{Income}{Median household income in 1998 dollars}
\item{Earnings}{Average hourly earnings of production workers in
manufacturing}
\item{Heart}{Deaths by heart disease per 100000
population}
\item{Vehicles}{Deaths by motor vehicle accidents per
100000 population}
 \item{Homeowners}{Home ownership rate}
}
}
\source{
United States Census Bureau \url{https://www.census.gov/}
}
\description{
Census data on the 50 states from 2003.
}
\keyword{datasets}
