% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CarletonStats-package.R
\docType{package}
\name{CarletonStats-package}
\alias{CarletonStats}
\alias{CarletonStats-package}
\title{CarletonStats: Functions for Statistics Classes at Carleton College}
\description{
Includes commands for bootstrapping and permutation tests, a command for created grouped bar plots, and a demo of the quantile-normal plot for data drawn from different distributions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/aloy/CarletonStats}
  \item Report bugs at \url{https://github.com/aloy/CarletonStats/issues}
}

}
\author{
\strong{Maintainer}: Adam Loy \email{aloy@carleton.edu} (\href{https://orcid.org/0000-0002-5780-4611}{ORCID})

Authors:
\itemize{
  \item Laura Chihara \email{flchihara@carleton.edu}
}

}
\keyword{internal}
