% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.R
\name{confint.carlboot}
\alias{confint.carlboot}
\title{Calculate a CI from a \code{carlboot} object}
\usage{
\method{confint}{carlboot}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{The carlboot object to print.}

\item{parm}{not used in CarletonStats, just for generic consistency}

\item{level}{the confidence level}

\item{...}{not used}
}
\description{
Calculate percentile confidence intervals for a \code{carlboot} object.
}
