% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.carlboot}
\alias{plot.carlboot}
\title{Plot the bootstrap distribution in \code{carlboot} object}
\usage{
\method{plot}{carlboot}(x, bins = 15, size = 5, ...)
}
\arguments{
\item{x}{The carlboot object to print.}

\item{bins}{number of bins in histogram.}

\item{size}{size of points.}

\item{...}{not used}
}
\description{
Plot the bootstrap distribution returned as a \code{carlboot} object.
}
