% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.carlboot}
\alias{summary.carlboot}
\alias{summary.carlperm}
\title{Print a summary of an \code{carlboot} object}
\usage{
\method{summary}{carlboot}(object, ...)

\method{summary}{carlperm}(object, ...)
}
\arguments{
\item{object}{The carlboot object to print.}

\item{...}{not used}
}
\description{
Print summary statistics and confidence intervals, if desired, for an \code{lmeresamp} object.
}
\examples{
boot_dist <- boot(ToothGrowth$len, ToothGrowth$supp, B = 1000)
summary(boot_dist)
perm_dist <- permTest(states03$ViolentCrime, states03$DeathPenalty, B = 999)
summary(perm_dist)
}
